backend/requirements.txt
# Contenido de backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
opencv-python